Implement TMA loads on views
