+++ crates/cubecl-std/Cargo.toml
@@ -21,5 +21,6 @@ export_tests = []
 cubecl-core = { path = "../cubecl-core", version = "0.7.0", default-features = false }
 cubecl-runtime = { path = "../cubecl-runtime", version = "0.7.0", default-features = false }
 half.workspace = true
+paste = { workspace = true }
 serde = { workspace = true }
 variadics_please = { workspace = true }
