+++ crates/cubecl-std/src/tensor/view/operations/mod.rs
@@ -1,0 +1,10 @@
+mod array;
+mod base;
+mod slice;
+mod tensor_map;
+mod view;
+mod virtual_tensor;
+mod virtual_view;
+
+pub use base::*;
+pub use virtual_view::*;
